rename communication_op to communication
