+++ fastdeploy/model_executor/graph_optimization/cudagraph_piecewise_backend.py
@@ -22,6 +22,7 @@ from paddle.device.cuda import graphs
 
 from fastdeploy.config import FDConfig
 from fastdeploy.utils import get_logger
+from fastdeploy.distributed.communication import capture_custom_allreduce
 
 logger = get_logger("cudagrpah_piecewise_backend", "cudagraph_piecewise_backend.log")
 
@@ -109,8 +110,7 @@ class CudaGraphPiecewiseBackend:
             paddle.device.synchronize()
 
             # Capture
-            from fastdeploy.distributed.communication_op import graph_capture
-            with graph_capture():
+            with capture_custom_allreduce():
                 new_grpah.capture_begin()
                 output = entry.runnable(**kwargs)
                 new_grpah.capture_end()
