+++ fastdeploy/model_executor/layers/backends/gcu/moe/fused_moe_method_gcu_backend.py
@@ -190,7 +190,7 @@ class GCUFusedMoeMethod(MoEMethodBase):
         fused_moe_out = fused_moe_out.reshape_([token_num, hidden_size])
 
         if layer.tp_size > 1:
-            from fastdeploy.distributed.communication_op import (
+            from fastdeploy.distributed.communication import (
                 tensor_model_parallel_all_reduce,
             )
