+++ fastdeploy/worker/gpu_worker.py
@@ -62,7 +62,7 @@ class GpuWorker(WorkerBase):
             gc.collect()
             paddle.device.cuda.empty_cache()
             if self.parallel_config.enable_custom_all_reduce:
-                from fastdeploy.distributed.communication_op import use_custom_allreduce
+                from fastdeploy.distributed.communication import use_custom_allreduce
 
                 use_custom_allreduce()
         else:
